Fixed pandas eq() and sped up hash
The equals method was not implemented as was therefore falling back to
the equals operator which is not defined for dataframes.  Now we,
correctly, use df.equals(other).
Hash has been changed to use a hash of the numpy array bytes
representation of a DF.  This is much faster and should give good
results.